@@ -13,18 +13,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/forms": "^7.2.2",
-    "@angular/http": "^7.2.2",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/router": "^7.2.2",
-    "@ionic-native/android-permissions": "^5.17.1",
-    "@ionic-native/background-mode": "^5.17.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "@ionic-native/android-permissions": "^5.8.0",
+    "@ionic-native/background-mode": "^5.8.0",
     "@ionic-native/camera": "^5.17.1",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/date-picker": "^5.17.1",
@@ -82,45 +83,48 @@
     "cordova-plugin-wifiwizard2": "git+https://github.com/tripflex/wifiwizard2.git",
     "cordova-sqlite-storage": "^3.4.0",
     "cordova-support-android-plugin": "1.0.1",
-    "cordova-support-google-services": "1.3.2",
+    "cordova-support-google-services": "^1.4.1",
     "core-js": "^2.5.4",
     "date-fns": "^1.30.1",
+    "es6-promise-plugin": "^4.1.0",
     "hammerjs": "^2.0.8",
+    "ion-bottom-drawer": "^2.0.0",
     "jspdf": "^1.5.3",
     "ng2-avatar": "^2.3.1",
     "ng2-img-max": "^2.1.18",
+    "ngx-barcode": "^0.2.4",
     "plugin": "^0.3.3",
     "rxjs": "~6.3.3",
     "rxjs-compat": "^6.4.0",
     "tslib": "^1.9.0",
+    "wifiwizard2": "^3.1.1",
     "xlsx": "^0.14.1",
     "zone.js": "~0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.12.4",
-    "@angular-devkit/build-angular": "0.12.4",
-    "@angular-devkit/core": "~7.3.8",
-    "@angular-devkit/schematics": "~7.3.8",
-    "@angular/cli": "~7.2.3",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
-    "@ionic/angular-toolkit": "~1.4.0",
-    "@types/node": "~12.0.0",
+    "@angular-devkit/build-angular": "~0.803.20",
+    "@angular/cli": "~8.3.23",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@ionic/angular-toolkit": "^2.1.1",
+    "@ionic/app-scripts": "^3.2.4",
+    "@ionic/lab": "2.0.6",
+    "@types/core-js": "^2.5.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
+    "@types/node": "^10.17.9",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
+    "karma": "~3.1.4",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.4.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.17.0",
-    "typescript": "~3.1.6"
+    "ts-node": "~8.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   },
   "description": "An Ionic project",
   "cordova": {
