@@ -7,7 +7,7 @@ export const environment = {
 // export const DOMAIN_URL = 'http://104.155.139.135:8080';
 
 // production from mobile
-// export const DOMAIN_URL = "http://104.198.175.198";
+export const DOMAIN_URL = "http://104.198.175.198";
 
 // local from mobile
-export const DOMAIN_URL = 'http://192.168.1.28:9090';
+// export const DOMAIN_URL = 'http://192.168.1.28:9090';
